requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.4
pydantic==2.7.1
python-dotenv==1.0.1